+++ rust/crates/astra-turn-core/src/lib.rs
@@ -165,7 +165,9 @@ pub mod headless_types;
 pub mod result_quality {
     pub use astra_turn_types::{ResultQuality, classify_result, quality_feedback};
 }
+pub mod agentic_prepare_payload;
 pub mod routing_engine;
 pub mod tool_registry_plugin;
+pub mod tool_registry_selection_edge_hints;
 pub mod turn_guard;
 pub mod ws_approval_gate;
